fix: 修复拉取 types 行为不对
