--- registry/new-york/blocks/index-table/types/index-table.ts
@@ -1,30 +1,0 @@
-import { type OrderDirection } from "@/registry/new-york/blocks/index-table/types/order-direction";
-
-export interface IndexTableOrder<OrderField> {
-  field: OrderField;
-  direction: OrderDirection;
-}
-
-export interface IndexTableEdge<Node> {
-  node: Node;
-  cursor: string;
-}
-
-export interface IndexTablePageInfo {
-  endCursor?: string | null;
-  hasNextPage: boolean;
-  hasPreviousPage: boolean;
-  startCursor?: string | null;
-}
-
-export interface IndexTablePagination {
-  first?: number;
-  last?: number;
-  before?: string | null;
-  after?: string | null;
-}
-
-export interface IndexTableValue<OrderField> extends IndexTablePagination {
-  query?: string;
-  orderBy?: IndexTableOrder<OrderField>;
-}
