--- registry/new-york/blocks/index-table/types/order-direction.ts
@@ -1,4 +1,0 @@
-export enum OrderDirection {
-  ASC = "ASC",
-  DESC = "DESC",
-}
